Add files via upload
added script to scrape egauges using xml and urllib2 which formed a new database named full_list.csv which houses the name of the egauge and the earliest timestamp solar data is available. This allows for the collection if exactly the right amount of weather data to match the solar data. That way there is no time wasted gathering data that will be cleaned away anyway.